+++ src/index.ts
@@ -35,7 +35,6 @@ export {
 } from './theme';
 export { Card } from './Card';
 export { Hero } from './Hero';
-export { Input } from './Form';
 export { Modal } from './Modal';
-export { Toggle } from './Toggle';
+export { Input, Toggle, Radio } from './Form';
 export * from './Icon';
